added validation on obstacleEventData with tests

+++ backend/src/business_logic_layer/obstacle_event_manager.js
@@ -1,17 +1,22 @@
+/* eslint-disable no-throw-literal */
 const { OBSTACLE_EVENT } = require('../presentation_layer/socketio/command_types')
 const fs = require('fs')
 
-module.exports = ({ obstacleEventRepository, uploadImage, SocketIOServer }) => {
+module.exports = ({ obstacleEventRepository, uploadImage, SocketIOServer, dtoValidator }) => {
     return {
         handleObstacleEvent: async (obstacleEventData) => {
-            // TODO: Validate incoming data
-            const { tmpImageFilePath, x, y } = obstacleEventData
+            obstacleEventData.x = Number(obstacleEventData.x)
+            obstacleEventData.y = Number(obstacleEventData.y)
+            const validationResult = dtoValidator.validateObstacleEventDto(obstacleEventData)
+            if (!validationResult) throw "Validation Error"
+
             // TODO: Correct x and y values to get reall position on map
+            const { tmpImageFilePath, x, y } = obstacleEventData
+
             try {
                 // [Du Won] Store image in Cloud Storage
                 const imageUrl = await uploadImage(tmpImageFilePath)
 
-                // TODO: Delete tmp image file
                 fs.unlink(tmpImageFilePath, (e) => {
                     if (e) {
                         // TODO: report that was unable to delete tmp file
